+++ CHANGES.md
@@ -1,4 +1,10 @@
-# v0.4.0
+# v1.0.2
+* Add mypy typing hint for non-tensorflow codes
+
+# v1.0.1
+* New find_points_in_spheres algorithm in pymatgen for graph construction 
+
+# v1.0.0
 * Tensorflow 2.0 version.
 
 # v0.3.4